moved css files in style directory
